scripts/requirements.txt
openpyxl>=3.0.0
python-docx>=0.8.11
markdown>=3.0
beautifulsoup4>=4.0
openai>=1.0.0
anthropic>=0.39.0
